requirements.txt
# Core Web Framework
streamlit==1.31.0
streamlit-option-menu==0.3.12
streamlit-lottie==0.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Data Processing
pandas==2.1.4
numpy<2.0
openpyxl==3.1.2

# Machine Learning & Statistics
scikit-learn==1.4.0
prophet==1.1.5
statsmodels==0.14.1
shap==0.44.1

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
reportlab==4.0.7
kaleido==0.2.1

# Additional ML Tools
xgboost==2.0.3
scipy==1.12.0
